requirements.txt
streamlit
pandas
numpy
seaborn
plotly
fpdf2
matplotlib